requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtpd==1.4.6
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
atpublic==5.0
attrs==24.2.0
billiard==4.2.1
boto3==1.35.70
botocore==1.35.70
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1.3
django-appconf==1.0.6
django-celery-email==3.0.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.5.0
idna==3.10
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
jmespath==1.0.1
kombu==5.4.2
matplotlib-inline==0.1.7
multidict==6.1.0
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.49
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.2
traitlets==5.14.3
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.2
